pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "prosper0"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "cryptography>=41.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sse-starlette>=1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
    "respx>=0.20",
]

[project.scripts]
prosper0-api = "stack.api:serve"

[tool.setuptools.packages.find]
include = ["stack*", "transparency*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
